rename vm/contract folder to vm/program
Rename contract folder to program. Move program and signature to
root directory.
